webconsole/backend/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-spanner==0.25.0
google-cloud-pubsub==0.27.0
google-cloud-storage==1.3.2
google-api-python-client==1.6.3
google-auth-httplib2==0.0.2
